requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
black==25.1.0
click==8.1.8
confuse==2.0.1
flake8==7.1.2
frozenlist==1.5.0
idna==3.10
isort==6.0.0
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.3.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyYAML==6.0.2
tomli==2.2.1
typing_extensions==4.12.2
yarl==1.18.3
